+++ packages/headless/src/case-assist.index.ts
@@ -53,17 +53,38 @@ export {buildCaseField} from './controllers/case-field/headless-case-field';
 export type {CaseFieldSuggestion} from './features/case-field/case-field-state';
 
 export type {
-  QuickviewCore,
-  QuickviewCoreOptions,
-  QuickviewCoreProps,
-  QuickviewCoreState,
+  Quickview,
+  QuickviewOptions,
+  QuickviewProps,
+  QuickviewState,
 } from './controllers/core/quickview/headless-core-quickview';
+
+export type {
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  Quickview as QuickviewCore,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuickviewOptions as QuickviewCoreOptions,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuickviewProps as QuickviewCoreProps,
+  /**
+   * @deprecated This is an internal controller that will be removed in the next version.
+   */
+  QuickviewState as QuickviewCoreState,
+} from './controllers/core/quickview/headless-core-quickview';
+
 export type {
   CaseAssistQuickviewState,
   CaseAssistQuickview,
   CaseAssistQuickviewOptions,
   CaseAssistQuickviewProps,
 } from './controllers/quickview/case-assist-headless-quickview';
+export {buildCaseAssistQuickview as buildQuickview} from './controllers/quickview/case-assist-headless-quickview';
 export {buildCaseAssistQuickview} from './controllers/quickview/case-assist-headless-quickview';
 
 export type {Result} from './api/search/search/result';
@@ -77,6 +98,8 @@ export type {
   DocumentSuggestionListState,
 } from './controllers/document-suggestion-list/headless-document-suggestion-list';
 export {buildDocumentSuggestionList} from './controllers/document-suggestion-list/headless-document-suggestion-list';
+
+export {buildCaseAssistInteractiveResult as buildInteractiveResult} from './controllers/document-suggestion-list/case-assist-headless-interactive-result';
 
 /**
  * @deprecated