+++ packages/headless/src/controllers/product-listing/category-facet/headless-product-listing-category-facet.ts
@@ -26,8 +26,6 @@ import {
 } from '../../../app/reducers';
 import {loadReducerError} from '../../../utils/errors';
 import {
-  CoreCategoryFacet,
-  CoreCategoryFacetState,
   buildCoreCategoryFacet,
   CategoryFacet,
   CategoryFacetProps,
@@ -35,14 +33,14 @@ import {
   CategoryFacetSearch,
   CategoryFacetSearchState,
   CategoryFacetSearchResult,
+  CoreCategoryFacet,
+  CoreCategoryFacetState,
 } from '../../core/facets/category-facet/headless-core-category-facet';
 import {fetchProductListing} from '../../../features/product-listing/product-listing-actions';
 import {ProductListingEngine} from '../../../app/product-listing-engine/product-listing-engine';
 import {buildCategoryFacetSearch} from './headless-product-listing-category-facet-search';
 
 export type {
-  CoreCategoryFacet,
-  CoreCategoryFacetState,
   CategoryFacetValue,
   CategoryFacetOptions,
   CategoryFacetSearchOptions,
@@ -52,6 +50,8 @@ export type {
   CategoryFacetSearch,
   CategoryFacetSearchState,
   CategoryFacetSearchResult,
+  CoreCategoryFacet,
+  CoreCategoryFacetState,
 };
 
 /**