+++ packages/quantic/cypress/integration/breadcrumb-manager/breadcrumb-manager.cypress.ts
@@ -11,7 +11,7 @@ describe('quantic-breadcrumb-manager', () => {
   const numericField = 'ytlikecount';
   const dateField = 'date';
   const categoryField = 'geographicalhierarchy';
-  const clearActionName = 'Clear';
+  const clearActionName = 'Clear filter';
 
   interface BreadcrumbOptions {
     categoryDivider: string;
@@ -101,8 +101,8 @@ describe('quantic-breadcrumb-manager', () => {
         Expect.dateFacetBreadcrumb.firstBreadcrumbValueLabelContains(
           timeframeLabel1
         );
-        Expect.dateFacetBreadcrumb.firstBreadcrumbAltTextContains(
-          `${dateFacetName} ${clearActionName}`
+        Expect.dateFacetBreadcrumb.firstBreadcrumbAltTextEq(
+          `${dateFacetName} ${timeframeLabel1} ${clearActionName}`
         );
 
         Actions.timeframeFacet.selectValue(timeframeLabel2);
@@ -140,8 +140,8 @@ describe('quantic-breadcrumb-manager', () => {
         Expect.categoryFacetBreadcrumb.firstBreadcrumbValueLabelContains(
           path[0]
         );
-        Expect.categoryFacetBreadcrumb.firstBreadcrumbAltTextContains(
-          `${categoryFacetName} ${clearActionName}`
+        Expect.categoryFacetBreadcrumb.firstBreadcrumbAltTextEq(
+          `${categoryFacetName} ${path[0]} ${clearActionName}`
         );
 
         Actions.categoryFacet.selectChildValue(path[1]);
@@ -235,8 +235,8 @@ describe('quantic-breadcrumb-manager', () => {
         Expect.dateFacetBreadcrumb.firstBreadcrumbValueLabelContains(
           'Past 6 months'
         );
-        Expect.dateFacetBreadcrumb.firstBreadcrumbAltTextContains(
-          `${dateFacetName} ${clearActionName}`
+        Expect.dateFacetBreadcrumb.firstBreadcrumbAltTextEq(
+          `${dateFacetName} Past 6 months ${clearActionName}`
         );
         Expect.categoryFacetBreadcrumb.displayBreadcrumb(true);
         Expect.categoryFacetBreadcrumb.displayLabel(true);