+++ packages/quantic/force-app/main/default/lwc/quanticModal/quanticModal.js
@@ -4,6 +4,7 @@ import {LightningElement, api} from 'lwc';
  * The `QuanticModal` is a container component that displays slotted content in a modal. This component handles the animation logic, exposes methods to open and close the modal, offers the option to open the modal in full screen or just to cover the search interface, and exposes a set of slots to fully customize the modal content.
  *
  * @category Search
+ * @category Insight Panel
  * @example
  * <c-quantic-modal full-screen animation="slideToLeft">
  *   <div slot="header">Modal Header</div>
@@ -14,8 +15,8 @@ import {LightningElement, api} from 'lwc';
 export default class QuanticModal extends LightningElement {
   animations = {
     slideToLeft: 'slideToLeft',
-    slideToTop: 'slideToTop'
-  }
+    slideToTop: 'slideToTop',
+  };
 
   /**
    * Indicates whether the modal will be opened in full screen.
@@ -42,10 +43,13 @@ export default class QuanticModal extends LightningElement {
   }
 
   validateProps() {
-    if (this.animation !== this.animations.slideToLeft && this.animation !== this.animations.slideToTop) {
+    if (
+      this.animation !== this.animations.slideToLeft &&
+      this.animation !== this.animations.slideToTop
+    ) {
       this.renderingError = `"${this.animation}" is an invalid value for the animation property. animation can only be set to "${this.animations.slideToTop}" or "${this.animations.slideToLeft}".`;
     }
-    if ( typeof(this.fullScreen)!== 'boolean' ){
+    if (typeof this.fullScreen !== 'boolean') {
       this.renderingError = `"${this.fullScreen}" is an invalid value for the full-screen property. full-screen can only be set to a boolean value.`;
     }
   }