+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.26...@coveo/atomic-react@1.16.0) (2022-07-28)
+
+
+### Features
+
+* **atomic,headless:** add result list and result template to insight ([#2197](https://github.com/coveo/ui-kit/issues/2197)) ([767b242](https://github.com/coveo/ui-kit/commit/767b24200e8fc43b64fb85075a97b574de5bc9ba))
+
+
+
+
+
+## [1.15.26](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.25...@coveo/atomic-react@1.15.26) (2022-07-28)
+
+
+### Bug Fixes
+
+* **atomic:** pass rendering function down to instant result atomic-re… ([#2257](https://github.com/coveo/ui-kit/issues/2257)) ([03336d6](https://github.com/coveo/ui-kit/commit/03336d6d2fc07561aa8dc88e2e2b6c99228d9315))
+
+
+
+
+
+## [1.15.25](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.24...@coveo/atomic-react@1.15.25) (2022-07-27)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.24](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.23...@coveo/atomic-react@1.15.24) (2022-07-27)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.23](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.22...@coveo/atomic-react@1.15.23) (2022-07-26)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [1.15.22](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.21...@coveo/atomic-react@1.15.22) (2022-07-26)
 
 **Note:** Version bump only for package @coveo/atomic-react